tetragon: Make TestUsdtArgs amd64 only
We only support parsing of usdt arguments for x86 arch,
so we need to run the test only on x86.
