+++ pkg/sensors/tracing/usdt_test.go
@@ -19,7 +19,6 @@ import (
 	"github.com/cilium/tetragon/pkg/bpf"
 	"github.com/cilium/tetragon/pkg/config"
 	"github.com/cilium/tetragon/pkg/jsonchecker"
-	lc "github.com/cilium/tetragon/pkg/matchers/listmatcher"
 	sm "github.com/cilium/tetragon/pkg/matchers/stringmatcher"
 	"github.com/cilium/tetragon/pkg/observer/observertesthelper"
 	"github.com/cilium/tetragon/pkg/sensors"
@@ -137,161 +136,6 @@ spec:
 		WithName(sm.Full("usdt0"))
 
 	checker := ec.NewUnorderedEventChecker(upChecker)
-
-	var doneWG, readyWG sync.WaitGroup
-	defer doneWG.Wait()
-
-	ctx, cancel := context.WithTimeout(context.Background(), tus.Conf().CmdWaitTime)
-	defer cancel()
-
-	obs, err := observertesthelper.GetDefaultObserverWithFile(t, ctx, testConfigFile, tus.Conf().TetragonLib, observertesthelper.WithMyPid())
-	if err != nil {
-		t.Fatalf("GetDefaultObserverWithFile error: %s", err)
-	}
-	observertesthelper.LoopEvents(ctx, t, &doneWG, &readyWG, obs)
-	readyWG.Wait()
-
-	if err := exec.Command(usdt).Run(); err != nil {
-		t.Fatalf("Failed to execute test binary: %s\n", err)
-	}
-
-	err = jsonchecker.JsonTestCheck(t, checker)
-	require.NoError(t, err)
-}
-
-func TestUsdtArgs(t *testing.T) {
-	if !config.EnableLargeProgs() || !bpf.HasUprobeRefCtrOffset() {
-		t.Skip("Need 5.3 or newer kernel for usdt and uprobe ref_ctr_off support for this test.")
-	}
-
-	usdt := testutils.RepoRootPath("contrib/tester-progs/usdt")
-	usdtHook := `
-apiVersion: cilium.io/v1alpha1
-kind: TracingPolicy
-metadata:
-  name: "usdts"
-spec:
-  usdts:
-  - path: "` + usdt + `"
-    provider: "test"
-    name: "usdt0"
-  - path: "` + usdt + `"
-    provider: "test"
-    name: "usdt3"
-    args:
-    - index: 0
-      type: "int32"
-    - index: 1
-      type: "int64"
-    - index: 2
-      type: "uint64"
-  - path: "` + usdt + `"
-    provider: "test"
-    name: "usdt12"
-    args:
-    - index: 0
-      type: "int32"
-    - index: 1
-      type: "int32"
-    - index: 2
-      type: "int64"
-    - index: 3
-      type: "int64"
-    - index: 4
-      type: "int32"
-  - path: "` + usdt + `"
-    provider: "test"
-    name: "usdt12"
-    args:
-    - index: 5
-      type: "int64"
-    - index: 6
-      type: "uint64"
-    - index: 7
-      type: "uint64"
-    - index: 8
-      type: "int32"
-    - index: 9
-      type: "int32"
-  - path: "` + usdt + `"
-    provider: "test"
-    name: "usdt12"
-    args:
-    - index: 10
-      type: "int32"
-    - index: 11
-      type: "int32"
-`
-
-	usdtConfigHook := []byte(usdtHook)
-	err := os.WriteFile(testConfigFile, usdtConfigHook, 0644)
-	if err != nil {
-		t.Fatalf("writeFile(%s): err %s", testConfigFile, err)
-	}
-
-	upChecker_0 := ec.NewProcessUsdtChecker("USDT_GENERIC_0").
-		WithProcess(ec.NewProcessChecker().
-			WithBinary(sm.Full(usdt))).
-		WithProvider(sm.Full("test")).
-		WithName(sm.Full("usdt0"))
-
-	upChecker_3 := ec.NewProcessUsdtChecker("USDT_GENERIC_3").
-		WithProcess(ec.NewProcessChecker().
-			WithBinary(sm.Full(usdt))).
-		WithProvider(sm.Full("test")).
-		WithName(sm.Full("usdt3")).
-		WithArgs(ec.NewKprobeArgumentListMatcher().
-			WithOperator(lc.Ordered).
-			WithValues(
-				ec.NewKprobeArgumentChecker().WithIntArg(1),
-				ec.NewKprobeArgumentChecker().WithLongArg(42),
-				ec.NewKprobeArgumentChecker().WithSizeArg(0xdeadbeef),
-			))
-
-	upChecker_12_1 := ec.NewProcessUsdtChecker("USDT_GENERIC_12_1").
-		WithProcess(ec.NewProcessChecker().
-			WithBinary(sm.Full(usdt))).
-		WithProvider(sm.Full("test")).
-		WithName(sm.Full("usdt12")).
-		WithArgs(ec.NewKprobeArgumentListMatcher().
-			WithOperator(lc.Ordered).
-			WithValues(
-				ec.NewKprobeArgumentChecker().WithIntArg(1),
-				ec.NewKprobeArgumentChecker().WithIntArg(2),
-				ec.NewKprobeArgumentChecker().WithLongArg(42),
-				ec.NewKprobeArgumentChecker().WithLongArg(43),
-				ec.NewKprobeArgumentChecker().WithIntArg(5),
-			))
-
-	upChecker_12_2 := ec.NewProcessUsdtChecker("USDT_GENERIC_12_2").
-		WithProcess(ec.NewProcessChecker().
-			WithBinary(sm.Full(usdt))).
-		WithProvider(sm.Full("test")).
-		WithName(sm.Full("usdt12")).
-		WithArgs(ec.NewKprobeArgumentListMatcher().
-			WithOperator(lc.Ordered).
-			WithValues(
-				ec.NewKprobeArgumentChecker().WithLongArg(6),
-				ec.NewKprobeArgumentChecker().WithSizeArg(0xdeadbeef),
-				ec.NewKprobeArgumentChecker().WithSizeArg(0xdeadbef7),
-				ec.NewKprobeArgumentChecker().WithIntArg(-9),
-				ec.NewKprobeArgumentChecker().WithIntArg(-2),
-			))
-
-	upChecker_12_3 := ec.NewProcessUsdtChecker("USDT_GENERIC_12_3").
-		WithProcess(ec.NewProcessChecker().
-			WithBinary(sm.Full(usdt))).
-		WithProvider(sm.Full("test")).
-		WithName(sm.Full("usdt12")).
-		WithArgs(ec.NewKprobeArgumentListMatcher().
-			WithOperator(lc.Ordered).
-			WithValues(
-				ec.NewKprobeArgumentChecker().WithIntArg(-3),
-				ec.NewKprobeArgumentChecker().WithIntArg(-127),
-			))
-
-	checker := ec.NewUnorderedEventChecker(upChecker_0, upChecker_3,
-		upChecker_12_1, upChecker_12_2, upChecker_12_3)
 
 	var doneWG, readyWG sync.WaitGroup
 	defer doneWG.Wait()
